requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
pandas==1.3.3
scikit-learn==0.24.2
numpy>=1.19.5,<2.0.0
tensorflow==2.6.0
typing-extensions==3.7.4
Werkzeug==2.1.0
